test(designTime): move discovery/execution tests in distinct directories
